dashboards/requirements.txt
# Kidjamo Streamlit Dashboards Requirements
streamlit>=1.28.0
boto3>=1.34.0
psycopg2-binary>=2.9.7
pandas>=1.5.0,<2.0.0
plotly>=5.17.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit-autorefresh>=0.0.1
streamlit-option-menu>=0.3.6
